requirements.txt
Django==1.6.1
Fabric==1.8.1
argparse==1.2.1
distribute==0.6.34
django-bower==4.8.1
django-braces==1.3.1
django-debug-toolbar==1.0.1
ecdsa==0.10
ipython==1.1.0
paramiko==1.12.1
pycrypto==2.6.1
six==1.5.2
sqlparse==0.1.10
wsgiref==0.1.2
